+++ src/main/kotlin/com/github/bucherfa/accessibilitylinter/mockLanguages/HandlebarsLanguage.kt
@@ -1,0 +1,6 @@
+package com.github.bucherfa.accessibilitylinter.mockLanguages
+
+import com.intellij.lang.Language
+
+// https://github.com/JetBrains/intellij-plugins/blob/master/handlebars/src/com/dmarcotte/handlebars/HbLanguage.java#L21
+class HandlebarsLanguage : Language("Handlebars")