+++ src/main/kotlin/com/github/bucherfa/accessibilitylinter/mockLanguages/VueJSLanguage.kt
@@ -1,0 +1,6 @@
+package com.github.bucherfa.accessibilitylinter.mockLanguages
+
+import com.intellij.lang.Language
+
+// https://github.com/JetBrains/intellij-plugins/blob/master/vuejs/src/org/jetbrains/vuejs/lang/html/VueLanguage.kt#L6
+class VueJSLanguage : Language("Vue")